requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
streamlit==1.38.0
pdfplumber==0.11.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
huggingface_hub==0.25.2
requests==2.32.3
numpy==1.26.4
